openfactory/fanoutlayer/asset_forwarder/requirements.txt
nats-py
confluent-kafka
